[MIG] mail_activity_team: Migration to 15.0
