+++ mail_activity_team/models/mail_activity.py
@@ -1,4 +1,4 @@
-# Copyright 2018 Eficent Business and IT Consulting Services, S.L.
+# Copyright 2018-22 ForgeFlow Business and IT Consulting Services, S.L.
 # License AGPL-3.0 or later (http://www.gnu.org/licenses/agpl.html).
 from odoo import SUPERUSER_ID, _, api, fields, models
 from odoo.exceptions import ValidationError
@@ -11,7 +11,7 @@ class MailActivity(models.Model):
         if not user_id:
             user_id = self.env.uid
         res_model = self.env.context.get("default_res_model")
-        model = self.env["ir.model"].search([("model", "=", res_model)], limit=1)
+        model = self.sudo().env["ir.model"].search([("model", "=", res_model)], limit=1)
         domain = [("member_ids", "in", [user_id])]
         if res_model:
             domain.extend(
@@ -74,6 +74,10 @@ class MailActivity(models.Model):
                 not in activity.team_id.with_context(active_test=False).member_ids
             ):
                 raise ValidationError(
-                    _("The assigned user %s is not member of the team %s.")
-                    % (activity.user_id.name, activity.team_id.name)
+                    _(
+                        "The assigned user %(user_name)s is "
+                        "not member of the team %(team_name)s.",
+                        user_name=activity.user_id.name,
+                        team_name=activity.team_id.name,
+                    )
                 )
