+++ mail_activity_team/readme/ROADMAP.rst
@@ -1,0 +1,1 @@
+* In next version systray.esm.js file should be changed as added in comments.
