--- mail_activity_team/views/assets_backend.xml
@@ -1,11 +1,0 @@
-<?xml version="1.0" encoding="utf-8" ?>
-<odoo>
-    <template id="assets_backend" name="mail assets" inherit_id="web.assets_backend">
-        <xpath expr="." position="inside">
-            <script
-                type="text/javascript"
-                src="/mail_activity_team/static/src/js/systray.js"
-            />
-        </xpath>
-    </template>
-</odoo>
